+++ cmd/command.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/giantswarm/devctl/v7/cmd/app"
 	"github.com/giantswarm/devctl/v7/cmd/completion"
+	"github.com/giantswarm/devctl/v7/cmd/deploy"
 	"github.com/giantswarm/devctl/v7/cmd/gen"
 	"github.com/giantswarm/devctl/v7/cmd/release"
 	"github.com/giantswarm/devctl/v7/cmd/replace"
@@ -138,6 +139,20 @@ func New(config Config) (*cobra.Command, error) {
 		}
 	}
 
+	var deployCmd *cobra.Command
+	{
+		c := deploy.Config{
+			Logger: logrus.StandardLogger(),
+			Stderr: config.Stderr,
+			Stdout: config.Stdout,
+		}
+
+		deployCmd, err = deploy.New(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
 	f := &flag{}
 
 	r := &runner{
@@ -161,6 +176,7 @@ func New(config Config) (*cobra.Command, error) {
 
 	c.AddCommand(appCmd)
 	c.AddCommand(completionCmd)
+	c.AddCommand(deployCmd)
 	c.AddCommand(genCmd)
 	c.AddCommand(releaseCmd)
 	c.AddCommand(replaceCmd)