+++ pkg/appstatus/error.go
@@ -1,0 +1,7 @@
+package appstatus
+
+import "github.com/giantswarm/microerror"
+
+var invalidConfigError = &microerror.Error{
+	Kind: "invalidConfigError",
+}